fix: revert suspense layout
