+++ src/app/page.tsx
@@ -1,30 +1,5 @@
-import { Suspense } from "react";
-import { notFound } from "next/navigation";
-import RoadmapListSkeleton from "@/features/roadmap/components/RoadmapListSkeleton";
-import RoadmapPaginationedList from "@/features/roadmap/components/RoadmapPaginationedList";
-import { GetRoadmapsParams } from "@/features/roadmap/type";
-import { FilterSearchParamsSchema } from "@/types";
+import { redirect } from "next/navigation";
 
-export default async function Home({
-  searchParams,
-}: {
-  searchParams: Promise<GetRoadmapsParams>;
-}) {
-  const safeParsedSearchParams = FilterSearchParamsSchema.safeParse(
-    await searchParams,
-  );
-
-  if (!safeParsedSearchParams.success) {
-    notFound();
-  }
-
-  return (
-    <Suspense fallback={<RoadmapListSkeleton />}>
-      <RoadmapPaginationedList
-        page={safeParsedSearchParams.data?.page}
-        categoryId={safeParsedSearchParams.data?.category}
-        keyword={safeParsedSearchParams.data?.keyword}
-      />
-    </Suspense>
-  );
+export default function RootPage() {
+  redirect("/roadmap");
 }
